docker build script takes build.env file and apply it to the container.

+++ deployment/docker/build.sh
@@ -19,9 +19,20 @@ function docker_rm() {
 
 trap "docker_rm" EXIT
 
+BUILD_ENV_PATH=${1:?"ERROR: env file is not given."}
+
+if [[ -n "${BUILD_ENV_PATH}" && ! -f "${BUILD_ENV_PATH}" ]]; then
+  echo "ERROR: Can't find the file: ${BUILD_ENV_PATH}" >&2
+  exit 1
+fi
+
+set -a
+. ${BUILD_ENV_PATH}
+set +a
+
 img_tag="openvnet/${BRANCH_NAME}"
 docker build -t "${img_tag}" - < "./deployment/docker/el7.Dockerfile"
-CID=$(docker run -d "${img_tag}")
+CID=$(docker run ${BUILD_ENV_PATH:+--env-file $BUILD_ENV_PATH} -d "${img_tag}")
 docker exec -t "${CID}" mkdir "/var/tmp/openvnet"
 docker cp . "${CID}:/var/tmp/openvnet"
 docker exec -t "${CID}" /bin/bash -c "cd openvnet; ./deployment/packagebuild/build_packages_vnet.sh"
